Remove all only 3D engine related stuff from codebase
