+++ OscilloscopeGraphics/OscilloscopeGraphics.h
@@ -1,9 +1,7 @@
 #ifndef __OSCILLOSCOPE_GRAPHICS__
 #define __OSCILLOSCOPE_GRAPHICS__
 
-#include "Camera.h"
-#include "Engine.h"
-#include "Mesh.h"
+#include "Clipper.h"
 #include "Renderer.h"
 #include "Server.h"
